% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addLeadingZeros}
\alias{addLeadingZeros}
\title{Add leading zeros to a numerical vector and convert it to string.}
\usage{
addLeadingZeros(num_vec)
}
\arguments{
\item{num_vec}{Numerical. Numerical vector.}
}
\value{
Character.
}
\description{
Add leading zeros to a numerical vector and convert it to string.
}
