% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CISBP_families}
\alias{CISBP_families}
\title{Dataframe of families of motifs from the CIS-BP database.}
\format{
Data frame
#' \describe{
  \item{motif}{Name of the motifs from CIS-BP database}
  \item{family}{TF family as determined by CISBP}
}
}
\usage{
CISBP_families
}
\description{
Dataframe of families of motifs from the CIS-BP database.
}
\keyword{datasets}
