% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CISBP_motifset}
\alias{CISBP_motifset}
\title{Motif set from the CIS-BP database.}
\format{
MEME format
}
\source{
\url{http://cisbp.ccbr.utoronto.ca/index.php}
}
\usage{
CISBP_motifset
}
\description{
Motif set from the CIS-BP database.
}
\keyword{datasets}
