% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_table_to_list.R
\name{cluster_table_to_list}
\alias{cluster_table_to_list}
\title{Convert a cluster table to a list of list object}
\usage{
cluster_table_to_list(cluster_table_file)
}
\arguments{
\item{cluster_table_file}{Character. Full path to output from the package PhotoperiodCluster}
}
\value{
List of list
}
\description{
Convert a cluster table to a list of list object
}
