% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concatenate_FASTA}
\alias{concatenate_FASTA}
\title{Concatenate a list of FASTA files}
\usage{
concatenate_FASTA(input_FASTA, output_FASTA)
}
\arguments{
\item{input_FASTA}{Character. Input FASTA file(s).}

\item{output_FASTA}{Character. Output FASTA file.}
}
\value{
None
}
\description{
Concatenate a list of FASTA files
}
