% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concatenate_gff3}
\alias{concatenate_gff3}
\title{Concatenate a list of GFF3 files}
\usage{
concatenate_gff3(input_gff3, output_gff3)
}
\arguments{
\item{input_gff3}{Character. Input GFF3 file(s).}

\item{output_gff3}{Character. Output GFF3 file.}
}
\value{
None
}
\description{
Concatenate a list of GFF3 files
}
