% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_TFfamily_colors}
\alias{df_TFfamily_colors}
\title{Dataframe of motif families from the CIS-BP database.}
\format{
Data frame
#' \describe{
  \item{family}{TF family as determined by CISBP}
  \item{Freq}{Frequency of transcription factor motifs in CISBP}
  \item{label}{Name of TF family that is to be displayed}
  \item{color}{RGB code of the family label}
}
}
\usage{
df_TFfamily_colors
}
\description{
Dataframe of motif families from the CIS-BP database.
}
\keyword{datasets}
