% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_motifs_fisher.R
\name{enrich_motifs_fisher}
\alias{enrich_motifs_fisher}
\title{Function to clean up .meme motif files. Remove duplicate names if there is any. Remove duplicate motifs if there is any.}
\usage{
enrich_motifs_fisher(
  qry_gff3,
  bkg_gff3,
  qry_fa,
  bkg_fa,
  alternative = "two.sided",
  motif_max_p = 1e-05
)
}
\arguments{
\item{qry_gff3}{Character vector of file paths of .gff3 files to be mapped on.}

\item{bkg_gff3}{Character vector of file paths of .gff3 files to be used as background.}

\item{qry_fa}{Character vector of file paths of .fa files to be mapped on.}

\item{bkg_fa}{Character vector of file paths of .fa files to be used as background.}

\item{alternative}{Character. Alternative hypothesis of the Fisher's test from the stats package.}

\item{motif_max_p}{Numerical. What value of p value to cap if very low p values were encountered.}
}
\value{
Data frame of the p value and fold enrichment.
}
\description{
Function to clean up .meme motif files. Remove duplicate names if there is any. Remove duplicate motifs if there is any.
}
