% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_result_summary.R
\name{enrich_result_summary}
\alias{enrich_result_summary}
\title{Processing of results from fisherEnrichMotif in a directory}
\usage{
enrich_result_summary(dir, max_pval = 0.05, outdir)
}
\arguments{
\item{dir}{Character. Input directory. Often the ouput directory of fisherEnrichMotif_inDir.}

\item{max_pval}{Numerical. p-value threshold above which the data will not be incoporated in the processed tables.}

\item{outdir}{Character. Output directory.}
}
\value{
None
}
\description{
Processing of results from fisherEnrichMotif in a directory
}
