% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes_by_rDEI.R
\name{filter_genes_by_rDEI}
\alias{filter_genes_by_rDEI}
\title{Function to filter gene IDs (TAIR IDs) based on daily expression integral}
\usage{
filter_genes_by_rDEI(geneID, rDEI_threshold = 1, rDEI_file = NULL)
}
\arguments{
\item{geneID}{Character vector of genes in TAIR ID}

\item{rDEI_threshold}{rDEI threshold}

\item{rDEI_file}{File of rDEI values}
}
\value{
Character vector of genes that pass the rDEI threshold
}
\description{
Function to filter gene IDs (TAIR IDs) based on daily expression integral
}
