% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_to_promoter_Biostring.R
\name{genes_to_promoter_Biostring}
\alias{genes_to_promoter_Biostring}
\title{Function to filter gene IDs (TAIR IDs) based on daily expression integral}
\usage{
genes_to_promoter_Biostring(
  gene_list,
  rDEI_threshold = 1,
  upstream = 1000,
  downstream = 500
)
}
\arguments{
\item{gene_list}{List of character vectors of gene IDs.}

\item{rDEI_threshold}{rDEI threshold for filtering}

\item{upstream}{Number of base pairs upstream of the transcription start site}

\item{downstream}{Number of base pairs downstream of the transcription start site}
}
\value{
Nothing returned
}
\description{
Function to filter gene IDs (TAIR IDs) based on daily expression integral
}
