% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{get_promoters}
\alias{get_promoters}
\title{Function to return Biostring object with input of gene models}
\usage{
get_promoters(gene_models, upstream, downstream)
}
\arguments{
\item{gene_models}{Character. Name of TAIR10 transcripts.}

\item{upstream}{Numerical. Number of base pair upstream of the transcription start site.}

\item{downstream}{Numerical. Number of base pair downstream of the transcription start site.}
}
\description{
Function to return Biostring object with input of gene models
}
