% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_motifs.R
\name{hclust_motifs}
\alias{hclust_motifs}
\title{Perform hierarchical clustering of motifs}
\usage{
hclust_motifs(
  motifs,
  compare_motifs_method = "PCC",
  hclust_method = "complete",
  min.position.ic = 0,
  min.mean.ic = 0.25,
  min.overlap = 6,
  normalize.scores = TRUE
)
}
\arguments{
\item{motifs}{Universalmotif object.}

\item{compare_motifs_method}{The similarity metric to be used between motifs according to the Universalmotifs package}

\item{hclust_method}{Method to perform hierarchical clustering.}

\item{min.position.ic}{The minimum IC between positions according to the Universalmotifs package}

\item{min.mean.ic}{The minimum mean IC according to the Universalmotifs package}

\item{min.overlap}{The minimum overlap between motifs according to the Universalmotifs package}

\item{normalize.scores}{Whether to normalize scores according to the Universalmotifs package}
}
\value{
hclust result
}
\description{
Perform hierarchical clustering of motifs
}
