% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_from_Fisher.R
\name{heatmap_from_Fisher}
\alias{heatmap_from_Fisher}
\title{Plot results of Fisher Exact Test as heat maps}
\usage{
heatmap_from_Fisher(
  matrix,
  hclust_out,
  cutree_out,
  motifs,
  style,
  remove_zoom = FALSE
)
}
\arguments{
\item{matrix}{Matrix. Numerical matrix to be plotted}

\item{hclust_out}{hclust object.}

\item{cutree_out}{cutree object.}

\item{motifs}{Universalmotifs object.}

\item{style}{Character. One of "FC", "PVAL" and "FDR".}

\item{remove_zoom}{whether to remove zoom box}
}
\value{
ComplexHeatmap object
}
\description{
Plot results of Fisher Exact Test as heat maps
}
