% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_motif_to_CISBP_family.R
\name{heatmap_motif_to_CISBP_family}
\alias{heatmap_motif_to_CISBP_family}
\title{Retrieve the motifs from a motif heatmap}
\usage{
heatmap_motif_to_CISBP_family(
  matrix,
  hclust_out,
  motifs,
  main,
  motif_similarity_cutoff = 0.9
)
}
\arguments{
\item{matrix}{input_matrix of the main heatmap}

\item{hclust_out}{hclust_out used to generate of the main heatmap}

\item{motifs}{Universalmotif object. The same one used to generate the main heatmap}

\item{main}{ComplexHeatmap object.}

\item{motif_similarity_cutoff}{similarity cutoff to determine closest motif in CISBP}
}
\value{
Character vector
}
\description{
Retrieve the motifs from a motif heatmap
}
