% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_TF_family.R
\name{heatmap_TF_family}
\alias{heatmap_TF_family}
\title{Perform hierarchical clustering of motifs}
\usage{
heatmap_TF_family(motif_family, custom_TF_family_scheme = NULL)
}
\arguments{
\item{motif_family}{Character.}

\item{custom_TF_family_scheme}{a custom naming scheme for the TF families as illustrated in df_TFfamily_colors}
}
\value{
List of ComplexHeatmap objects
}
\description{
Perform hierarchical clustering of motifs
}
