% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JASPAR_motifset}
\alias{JASPAR_motifset}
\title{Motif set from the JASPAR Plants database.}
\format{
MEME format
}
\usage{
JASPAR_motifset
}
\description{
Motif set from the JASPAR Plants database.
}
\keyword{datasets}
