% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_TAIR10_TxDb.R
\name{make_TAIR10_TxDb}
\alias{make_TAIR10_TxDb}
\title{Function to generate a TxDb object from the GenomicFeatures object for getting gene promoters.}
\usage{
make_TAIR10_TxDb()
}
\value{
Returns a TxDb object (from the GenomicFeatures package)
}
\description{
Function to generate a TxDb object from the GenomicFeatures object for getting gene promoters.
}
