% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_clean_up.R
\name{motif_clean_up}
\alias{motif_clean_up}
\title{Function to clean up .meme motif files. Remove duplicate names if there is any. Remove duplicate motifs if there is any.}
\usage{
motif_clean_up(
  input_files,
  keep_duplicates = FALSE,
  keep_same_name_motifs = TRUE,
  output_file = NULL
)
}
\arguments{
\item{input_files}{Character vector of file paths of .meme files to be cleaned up.}

\item{keep_duplicates}{Whether to keep motifs with the same positional weight matrices.}

\item{keep_same_name_motifs}{Whether to keep motifs with the same name.}

\item{output_file}{Whether to write a .meme motif file.}
}
\value{
A universalmotif object or nothing if output_file=TRUE.
}
\description{
Function to clean up .meme motif files. Remove duplicate names if there is any. Remove duplicate motifs if there is any.
}
