% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_to_CISBP_family.R
\name{motif_to_CISBP_family}
\alias{motif_to_CISBP_family}
\title{Search which family the most motif belongs to in CISBP}
\usage{
motif_to_CISBP_family(
  motifs,
  similarity_cutoff = 0.5,
  only_CISBP_motifs = FALSE
)
}
\arguments{
\item{motifs}{Universalmotif objects}

\item{similarity_cutoff}{Pearson's similarity cutoff for determining if a motif has a counterpart in CISBP}

\item{only_CISBP_motifs}{Whether the input consists of only motifs from CISBP and comparison is unnecessary}
}
\value{
Character
}
\description{
Search which family the most motif belongs to in CISBP
}
