% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_to_similar_CISBP.R
\name{motif_to_similar_CISBP}
\alias{motif_to_similar_CISBP}
\title{Find similar motifs against the CIS-BP motif database}
\usage{
motif_to_similar_CISBP(query_motifs)
}
\arguments{
\item{query_motifs}{Universalmotif objects}
}
\value{
Dataframe object
}
\description{
Find similar motifs against the CIS-BP motif database
}
