% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{photoperiodic_clusters}
\alias{photoperiodic_clusters}
\title{Photoperiodic clusters}
\format{
List of list of characters
}
\usage{
photoperiodic_clusters
}
\description{
Photoperiodic clusters
}
\keyword{datasets}
