% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pivotDFinDir}
\alias{pivotDFinDir}
\title{Merge all the tabular text files in a directory and perform an optional pivot}
\usage{
pivotDFinDir(
  input_dir,
  ID_col_name,
  value_col_name,
  sep = "\\t",
  skip_pivot = FALSE
)
}
\arguments{
\item{input_dir}{Character. Name of directory.}

\item{ID_col_name}{Character. Name of the column that contains the ID to be used as pivot.}

\item{value_col_name}{Character. Name of the column that contains the value to be used as pivot.}

\item{sep}{Character. Separator of the tabular data.}

\item{skip_pivot}{Logical. Whether the pivot should be performed.}
}
\value{
Data frame
}
\description{
Merge all the tabular text files in a directory and perform an optional pivot
}
