% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positional_bias.R
\name{plot_positional_bias}
\alias{plot_positional_bias}
\title{Plot the result from analyze_positional_bias}
\usage{
plot_positional_bias(df)
}
\arguments{
\item{df}{Output from the analyze_positional_bias function}
}
\value{
ggplot2 object
}
\description{
Plot the result from analyze_positional_bias
}
