% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_bias.R
\name{positional_bias}
\alias{positional_bias}
\title{Analyse positional bias of motifs in a set of DNA sequences}
\usage{
positional_bias(
  motifs,
  seqs,
  separate_motifs = TRUE,
  seq_names = NULL,
  pvalue = 1e-06,
  reverse_complement = TRUE,
  bandwidth = NULL
)
}
\arguments{
\item{motifs}{Universalmotif object or name of a .meme file with motifs}

\item{seqs}{List of Biostrings objects or list of file names of .fa file for DNA sequences}

\item{separate_motifs}{Logical. Whether individual motif should be analyzed separately}

\item{seq_names}{Names of the individual items in the list_of_seqs}

\item{pvalue}{Numerical. p-value for calling a motif peak}

\item{reverse_complement}{Logical. Whether the reverse complement is used to look for motifs}

\item{bandwidth}{Bandwidth of the universalmotif package}
}
\value{
Data frame
}
\description{
Analyse positional bias of motifs in a set of DNA sequences
}
