% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_bias_subfunc.R
\name{positional_bias_subfunc}
\alias{positional_bias_subfunc}
\title{Analyse positional bias of motifs in a set of DNA sequences}
\usage{
positional_bias_subfunc(
  motifs,
  list_of_seqs,
  seq_names = NULL,
  motif_name = "NA",
  p_value = 1e-06,
  reverse_complement = TRUE,
  bandwidth = NULL
)
}
\arguments{
\item{motifs}{Universalmotif object or name of a .meme file with motifs}

\item{list_of_seqs}{List of Biostrings objects or list of file names of .fa file for DNA sequences}

\item{seq_names}{Names of the individual items in the list_of_seqs}

\item{motif_name}{Character. Name of the motif}

\item{p_value}{Numerical. p-value for calling a motif peak}

\item{reverse_complement}{Logical. Whether the reverse complement is used to look for motifs}

\item{bandwidth}{Bandwidth of the universalmotif package}
}
\value{
Data frame
}
\description{
Analyse positional bias of motifs in a set of DNA sequences
}
