% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rDEI_summary}
\alias{rDEI_summary}
\title{TAIR10 gene models from Arabidopsis_thaliana.TAIR10.47.gtf}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{EQ_DEI}{daily expression integral at equinox photoperiod}
  \item{LD_DEI}{daily expression integral at long day photoperiod}
  \item{SD_DEI}{daily expression integral at short day photoperiod}
  \item{geneID}{}
  \item{log2_rDEI_LDEQ}{}
  \item{log2_rDEI_SDEQ}{}
  \item{log2_rDEI_SDLD}{}
  \item{rDEI_LDEQ}{}
  \item{rDEI_SDEQ}{}
  \item{rDEI_SDLD}{}

}
}
\usage{
rDEI_summary
}
\description{
TAIR10 gene models from Arabidopsis_thaliana.TAIR10.47.gtf
}
\keyword{datasets}
