% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Fisher_result.R
\name{read_Fisher_result}
\alias{read_Fisher_result}
\title{Retrieve results from Fisher's Exact Test}
\usage{
read_Fisher_result(in_dir, max_pval_cap = 100)
}
\arguments{
\item{in_dir}{Directory with the outputs of processFisherData}

\item{max_pval_cap}{Numerical. The maximum p value to display on a heatmap}
}
\value{
A list of mat_logp, mat_logfdr, mat_FC and motif_names
}
\description{
Retrieve results from Fisher's Exact Test
}
