% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_genes}
\alias{read_genes}
\title{Read a .txt file line by line and return a character vector.}
\usage{
read_genes(x)
}
\arguments{
\item{x}{Character. Name of a .txt file}
}
\value{
Character.
}
\description{
Read a .txt file line by line and return a character vector.
}
