% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_enrich.R
\name{run_enrich}
\alias{run_enrich}
\title{Function to perform motif enrichment from motif mapping result}
\usage{
run_enrich(dir, FASTA_files, GFF3_files)
}
\arguments{
\item{dir}{Character. Output directory.}

\item{FASTA_files}{Character vector of full paths to FASTA .fa files.}

\item{GFF3_files}{Character vector of full paths to GFF3 .gff3 files.}
}
\value{
None.
}
\description{
Function to perform motif enrichment from motif mapping result
}
