% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_heatmap.R
\name{run_heatmap}
\alias{run_heatmap}
\title{Function to plot motif enrichment results as heat maps}
\usage{
run_heatmap(
  dir,
  enrichment_dir,
  motif_MEME_file,
  enrich_pval_threshold = 0.01,
  cutree_height = 0.5,
  remove_zoom = FALSE
)
}
\arguments{
\item{dir}{Character. Output directory.}

\item{enrichment_dir}{Character. Full path to the run_enrich output directory.}

\item{motif_MEME_file}{Character. Full path to a motif .meme file.}

\item{enrich_pval_threshold}{Numerical. Default 1e-2. P value threshold to decide for an enriched motif.}

\item{cutree_height}{Numerical. Similarity value to cut the dendrogram for merging motifs.}

\item{remove_zoom}{whether to remove zoom box}
}
\value{
None.
}
\description{
Function to plot motif enrichment results as heat maps
}
