% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_input.R
\name{run_input}
\alias{run_input}
\title{Function to take in motifs and sequences as inputs}
\usage{
run_input(dir, motifs, sequences)
}
\arguments{
\item{dir}{Character. Output directory.}

\item{motifs}{Either a character vector of full paths to .meme motifs files, or a list of universalmotif objects.}

\item{sequences}{Either a character vector of full paths to .fa DNA sequence files, of a list of Biostring DNAStringSet / DNAString objects.}
}
\value{
None.
}
\description{
Function to take in motifs and sequences as inputs
}
