% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mapping.R
\name{run_mapping}
\alias{run_mapping}
\title{Function to perform motif mapping}
\usage{
run_mapping(dir, FASTA_files, motif_MEME_file, mapping_pval_threshold = 1e-04)
}
\arguments{
\item{dir}{Character. Output directory.}

\item{FASTA_files}{Character vector of full paths to FASTA .fa files.}

\item{motif_MEME_file}{Character. Full path to a motif .meme file.}

\item{mapping_pval_threshold}{Numerical. Default 1e-4. Threshold to determine whether a motif has been mapped to a stretch of sequence.}
}
\value{
None.
}
\description{
Function to perform motif mapping
}
