% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_master.R
\name{run_master}
\alias{run_master}
\title{Function to perform motif enrichment pipeline from motifs and sequences.}
\usage{
run_master(
  output_dir,
  motifs,
  sequences,
  skip_input = FALSE,
  skip_mapping = FALSE,
  skip_enrich = FALSE,
  skip_heatmap = FALSE
)
}
\arguments{
\item{output_dir}{Character. Output directory.}

\item{motifs}{Either a character vector of full paths to .meme motifs files, or a list of universalmotif objects.}

\item{sequences}{Either a character vector of full paths to .fa DNA sequence files, of a list of Biostring DNAStringSet / DNAString objects.}

\item{skip_input}{skip input}

\item{skip_mapping}{skip mapping}

\item{skip_enrich}{skip enrich}

\item{skip_heatmap}{skip heat map}
}
\value{
None.
}
\description{
Function to perform motif enrichment pipeline from motifs and sequences.
}
