% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Prism.R
\name{theme_Prism}
\alias{theme_Prism}
\title{Theme for imitating Prism}
\usage{
theme_Prism(base_size = 7, base_family = "Arial")
}
\arguments{
\item{base_size}{Numerical. Font size.}

\item{base_family}{Character. Font family.}
}
\value{
ggplot2 theme
}
\description{
Theme for imitating Prism
}
